+++ inc/data/services/PlacesService.php
@@ -29,12 +29,13 @@ class PlacesService {
         return $place;
     }
 
-    private static function buildDBFloor($db_floor, $pdo) {
-        $floor = Floor::__build($db_floor['ID'], $db_floor['NAME']);
-        $sqlplaces = 'SELECT * FROM PLACES WHERE FLOOR = "'
-                     . $db_floor['ID'] . '"';
-        foreach ($pdo->query($sqlplaces) as $db_place) {
-            $place = PlacesService::buildDBPlace($db_place);
+    private static function buildDBFloor($dbFloor, $pdo) {
+        $floor = Floor::__build($dbFloor['ID'], $dbFloor['NAME']);
+        $stmt = $pdo->prepare("SELECT * FROM PLACES WHERE FLOOR = :id");
+        $stmt->bindParam(":id", $dbFloor['ID']);
+        $stmt->execute();
+        while ($dbPlace = $stmt->fetch()) {
+            $place = PlacesService::buildDBPlace($dbPlace);
             $floor->addPlace($place);
         }
         return $floor;
@@ -44,8 +45,8 @@ class PlacesService {
         $floors = array();
         $pdo = PDOBuilder::getPDO();
         $sql = "SELECT * FROM FLOORS";
-        foreach ($pdo->query($sql) as $db_floor) {
-            $floor = PlacesService::buildDBFloor($db_floor, $pdo);
+        foreach ($pdo->query($sql) as $dbFloor) {
+            $floor = PlacesService::buildDBFloor($dbFloor, $pdo);
             $floors[] = $floor;
         }
         return $floors;
@@ -69,7 +70,7 @@ class PlacesService {
         $sql = "INSERT INTO FLOORS (ID, NAME, IMAGE) VALUES (:id, :name, :img)";
         $stmt = $pdo->prepare($sql);
         $stmt->bindParam(":id", $id, \PDO::PARAM_STR);
-        $stmt->bindParam(":name", $floor->name, \PDO::PARAM_STR);
+        $stmt->bindParam(":name", $floor->label, \PDO::PARAM_STR);
         $stmt->bindParam(":img",$img);
         if (!$stmt->execute()) {
             return false;
@@ -79,6 +80,11 @@ class PlacesService {
 
     static function deleteFloor($id) {
         $pdo = PDOBuilder::getPDO();
+        $stmt = $pdo->prepare("DELETE FROM PLACES WHERE FLOOR = :id");
+        $stmt->bindParam(":id", $id);
+        if ($stmt->execute() === false) {
+            return false;
+        }
         $stmt = $pdo->prepare("DELETE FROM FLOORS WHERE ID = :id");
         if ($stmt->execute(array(':id' => $id))) {
             return true;
@@ -95,21 +101,24 @@ class PlacesService {
         $sql .= "WHERE ID = :id";
         $stmt = $pdo->prepare($sql);
 
-        $stmt->bindParam(":name", $floor->name, \PDO::PARAM_STR);
+        $stmt->bindParam(":name", $floor->label, \PDO::PARAM_STR);
         if ($image !== "") {
             $stmt->bindParam(":image", $image);
         }
         $stmt->bindParam(":id", $floor->id, \PDO::PARAM_STR);
-
-        return $stmt->execute();
+        if ($stmt->execute() !== false) {
+            return true;
+        } else {
+            return false;
+        }
     }
 
     static function getAllPlaces() {
         $place = array();
         $pdo = PDOBuilder::getPDO();
         $sql = "SELECT * FROM PLACES";
-        foreach ($pdo->query($sql) as $db_place) {
-            $place = PlacesService::buildDBPlace($db_place, $pdo);
+        foreach ($pdo->query($sql) as $dbPlace) {
+            $place = PlacesService::buildDBPlace($dbPlace, $pdo);
             $places[] = $place;
         }
         return $places;
@@ -122,7 +131,7 @@ class PlacesService {
         $stmt->bindParam(":floor", $idFloor, \PDO::PARAM_STR);
         if ($stmt->execute(array(':floor' => $idFloor))) {
             while( $db_place = $stmt->fetch()) {
-                $place = PlacesService::buildDBPlace($db_place, $pdo);
+                $place = PlacesService::buildDBPlace($dbPlace, $pdo);
                 $places[] = $place;
             }
         }
@@ -155,10 +164,10 @@ class PlacesService {
                 . " WHERE ID = :id";
         $stmt = $pdo->prepare($sql);
 
-        $stmt->bindParam(":name", $place->name, \PDO::PARAM_STR);
+        $stmt->bindParam(":name", $place->label, \PDO::PARAM_STR);
         $stmt->bindParam(":x", $place->x, \PDO::PARAM_INT);
         $stmt->bindParam(":y", $place->y, \PDO::PARAM_INT);
-        $stmt->bindParam(":floor", $place->floor, \PDO::PARAM_STR);
+        $stmt->bindParam(":floor", $place->floorId, \PDO::PARAM_STR);
         $stmt->bindParam(":id", $place->id, \PDO::PARAM_STR);
 
         return $stmt->execute();
@@ -173,10 +182,10 @@ class PlacesService {
         $stmt = $pdo->prepare($sql);
 
         $stmt->bindParam(":id", $id, \PDO::PARAM_STR);
-        $stmt->bindParam(":name", $place->name, \PDO::PARAM_STR);
+        $stmt->bindParam(":name", $place->label, \PDO::PARAM_STR);
         $stmt->bindParam(":x", $place->x, \PDO::PARAM_INT);
         $stmt->bindParam(":y", $place->y, \PDO::PARAM_INT);
-        $stmt->bindParam(":floor", $place->floor, \PDO::PARAM_STR);
+        $stmt->bindParam(":floor", $place->floorId, \PDO::PARAM_STR);
 
         if (!$stmt->execute()) {
             return NULL;
